% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launches the EscolesOrbita app}
\usage{
launchApp()
}
\value{
shiny application object
}
\description{
launches the EscolesOrbita app
}
\examples{
\dontrun {launchApp()}

}
